requirements.txt
flask
werkzeug
pandas
PyPDF2
python-docx
scikit-learn
transformers
torch
sentencepiece
